refactor: move cluster package out of core (#1367)
