+++ pkg/cluster/clean.go
@@ -2,9 +2,9 @@ package cluster
 
 import (
 	"github.com/AlecAivazis/survey/v2"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/constants"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/services/resources"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/v1alpha"
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/constants"
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/services/resources"
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/v1alpha"
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/icon"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"k8s.io/apimachinery/pkg/api/errors"
