+++ pkg/cluster/kubeclient/kubeclients.go
@@ -2,9 +2,10 @@ package kubeclient
 
 import (
 	"fmt"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/v1alpha"
 	"os"
 	"path/filepath"
+
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/v1alpha"
 
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
