+++ pkg/cmd/cluster/bind/bind.go
@@ -2,10 +2,11 @@ package bind
 
 import (
 	"context"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/kubeclient"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/services/resources"
-	"github.com/redhat-developer/app-services-cli/pkg/core/cluster/v1alpha"
+
+	"github.com/redhat-developer/app-services-cli/pkg/cluster"
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/kubeclient"
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/services/resources"
+	"github.com/redhat-developer/app-services-cli/pkg/cluster/v1alpha"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/factory"
 	"github.com/redhat-developer/app-services-cli/pkg/core/config"
 	"github.com/redhat-developer/app-services-cli/pkg/core/connection"
